requirements.txt
Django==1.4
Pillow==1.7.7
South==0.7.6
argparse==1.2.1
cmsplugin-contact==1.0.0
cmsplugin-filer==0.9
distribute==0.6.28
django-classy-tags==0.3.4.1
django-cms==2.3.3
django-filer==0.9
django-mptt==0.5.2
django-polymorphic==0.2
django-reversion==1.6.3
django-sekizai==0.6.1
easy-thumbnails==1.1
html5lib==0.95
wsgiref==0.1.2
